src/main/java/com/fengs/service/impl/CategoryServiceImpl.java
package com.fengs.service.impl;

import com.fengs.dataobject.ProductCategory;
import com.fengs.repository.ProductCategoryRepository;
import com.fengs.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Created by fengs on 2018/1/27.
 */
@Service
public class CategoryServiceImpl  implements CategoryService{
    @Autowired
    private ProductCategoryRepository productCategoryRepository;
    @Override
    public ProductCategory findOne(Integer categoryId) {
    //  ProductCategory productCategory=  productCategoryRepository.findOne();
        return productCategoryRepository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryRepository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return productCategoryRepository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return productCategoryRepository.save(productCategory);
    }
}
